OO-133 : if basefile does not exist, delete should not throw an error! just return ok.
--HG--
branch : 80-patch

+++ src/main/java/org/olat/core/util/vfs/LocalFolderImpl.java
@@ -235,6 +235,9 @@ public class LocalFolderImpl extends LocalImpl implements VFSContainer {
 	 * @see org.olat.core.util.vfs.VFSItem#delete()
 	 */
 	public VFSStatus delete() {
+		if(!getBasefile().exists()){
+			return VFSConstants.YES;  // already non-existent
+		}
 		// we must empty the folders and subfolders first
 		for (Iterator<VFSItem> it_chd = getItems().iterator(); it_chd.hasNext();) {
 			it_chd.next().delete(); // TODO status
